@@ -1,7 +1,7 @@
 source "https://rubygems.org"
 git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-ruby '2.5.5'
+ruby '2.6.6'
 
 group :development, :test do
   gem "toml-rb"
@@ -12,9 +12,7 @@ group :development, :test do
   gem "rake"
   gem "parallel_tests"
   gem 'rspec-retry'
-  gem "netrc"
-  gem "git", github: "hone/ruby-git", branch: "master"
-  gem 'json', '~> 2.0.2'
+  gem 'json'
   gem 'ci-queue'
   gem 'redis'
 end
