@@ -13,7 +13,7 @@ module LanguagePack
     Instrument.instrument 'detect' do
       Dir.chdir(args.first)
 
-      pack = [ NoLockfile, Rails5, Rails42, Rails41, Rails4, Rails3, Rails2, Rack, Ruby ].detect do |klass|
+      pack = [ NoLockfile, Rails6, Rails5, Rails42, Rails41, Rails4, Rails3, Rails2, Rack, Ruby ].detect do |klass|
         klass.use?
       end
 
@@ -34,6 +34,8 @@ require "language_pack/helpers/stale_file_cleaner"
 require "language_pack/helpers/rake_runner"
 require "language_pack/helpers/rails_runner"
 require "language_pack/helpers/bundler_wrapper"
+require "language_pack/helpers/outdated_ruby_version"
+require "language_pack/helpers/download_presence"
 require "language_pack/installers/ruby_installer"
 require "language_pack/installers/heroku_ruby_installer"
 require "language_pack/installers/rbx_installer"
@@ -47,4 +49,5 @@ require "language_pack/rails4"
 require "language_pack/rails41"
 require "language_pack/rails42"
 require "language_pack/rails5"
+require "language_pack/rails6"
 require "language_pack/no_lockfile"
