@@ -5,12 +5,18 @@ require "language_pack"
 class LanguagePack::Cache
   # @param [String] path to the cache store
   def initialize(cache_path)
-    @cache_base = Pathname.new(cache_path)
+    if cache_path
+      @cache_base = Pathname.new(cache_path)
+    else
+      @cache_base = nil
+    end
   end
 
   # removes the the specified path from the cache
   # @param [String] relative path from the cache_base
   def clear(path)
+    return unless @cache_base
+
     target = (@cache_base + path)
     target.exist? && target.rmtree
   end
@@ -22,6 +28,8 @@ class LanguagePack::Cache
   # @param [String] path of contents to store. it will be stored using this a relative path from the cache_base.
   # @param [String] relative path to store the cache contents, if nil it will assume the from path
   def store(from, path = nil)
+    return unless @cache_base
+
     path ||= from
     clear path
     copy from, (@cache_base + path)
@@ -30,6 +38,8 @@ class LanguagePack::Cache
   # Adds file to cache without clearing the destination
   # Use LanguagePack::Cache#store to avoid accidental cache bloat
   def add(from, path = nil)
+    return unless @cache_base
+
     path ||= from
     copy from, (@cache_base + path)
   end
@@ -38,11 +48,15 @@ class LanguagePack::Cache
   # @param [String] relative path of the cache contents
   # @param [String] path of where to store it locally, if nil, assume same relative path as the cache contents
   def load(path, dest = nil)
+    return unless @cache_base
+
     dest ||= path
     copy (@cache_base + path), dest
   end
 
   def load_without_overwrite(path, dest=nil)
+    return unless @cache_base
+
     dest ||= path
     copy (@cache_base + path), dest, '-a -n'
   end
@@ -51,6 +65,8 @@ class LanguagePack::Cache
   # @param [String] source directory
   # @param [String] destination directory
   def copy(from, to, options='-a')
+    return unless @cache_base
+
     return false unless File.exist?(from)
     FileUtils.mkdir_p File.dirname(to)
     system("cp #{options} #{from}/. #{to}")
@@ -60,6 +76,8 @@ class LanguagePack::Cache
   # @param [String] source cache directory
   # @param [String] destination directory
   def cache_copy(from,to)
+    return unless @cache_base
+
     copy(@cache_base + from, @cache_base + to)
   end
 
@@ -67,6 +85,8 @@ class LanguagePack::Cache
   # @param [String] relative path of the cache contents
   # @param [Boolean] true if the path exists in the cache and false if otherwise
   def exists?(path)
+    return unless @cache_base
+
     File.exists?(@cache_base + path)
   end
 end
