@@ -11,15 +11,24 @@ class LanguagePack::Metadata
     end
   end
 
+  def [](key)
+    read(key)
+  end
+
+  def []=(key, value)
+    write(key, value)
+  end
+
   def read(key)
     full_key = "#{FOLDER}/#{key}"
-    File.read(full_key) if exists?(key)
+    File.read(full_key).strip if exists?(key)
   end
 
   def exists?(key)
     full_key = "#{FOLDER}/#{key}"
     File.exists?(full_key) && !Dir.exists?(full_key)
   end
+  alias_method :include?, :exists?
 
   def write(key, value, isave = true)
     FileUtils.mkdir_p(FOLDER)
@@ -27,9 +36,24 @@ class LanguagePack::Metadata
     full_key = "#{FOLDER}/#{key}"
     File.open(full_key, 'w') {|f| f.puts value }
     save if isave
+
+    return true
   end
 
-  def save
-    @cache ? @cache.add(FOLDER) : false
+  def touch(key)
+    write(key, "true")
+  end
+
+  def fetch(key)
+    return read(key) if exists?(key)
+
+    value = yield
+
+    write(key, value.to_s)
+    return value
+  end
+
+  def save(file = FOLDER)
+    @cache ? @cache.add(file) : false
   end
 end
