@@ -1,7 +1,7 @@
 require "language_pack"
-require "language_pack/base"
+require "language_pack/ruby"
 
-class LanguagePack::NoLockfile < LanguagePack::Base
+class LanguagePack::NoLockfile < LanguagePack::Ruby
   def self.use?
     !File.exists?("Gemfile.lock")
   end
