@@ -28,8 +28,8 @@ class LanguagePack::Rack < LanguagePack::Ruby
     instrument "rack.default_process_types" do
       # let's special case thin here if we detect it
       web_process = bundler.has_gem?("thin") ?
-        "bundle exec thin start -R config.ru -e $RACK_ENV -p $PORT" :
-        "bundle exec rackup config.ru -p $PORT"
+        "bundle exec thin start -R config.ru -e $RACK_ENV -p ${PORT:-5000}" :
+        "bundle exec rackup config.ru -p ${PORT:-5000}"
 
       super.merge({
         "web" => web_process
@@ -40,8 +40,8 @@ class LanguagePack::Rack < LanguagePack::Ruby
 private
 
   # sets up the profile.d script for this buildpack
-  def setup_profiled
-    super
+  def setup_profiled(*args)
+    super(*args)
     set_env_default "RACK_ENV", "production"
   end
 
