@@ -16,8 +16,8 @@ class LanguagePack::Rails2 < LanguagePack::Ruby
     end
   end
 
-  def initialize(build_path, cache_path=nil)
-    super(build_path, cache_path)
+  def initialize(*args)
+    super(*args)
     @rails_runner = LanguagePack::Helpers::RailsRunner.new
   end
 
@@ -45,8 +45,8 @@ class LanguagePack::Rails2 < LanguagePack::Ruby
   def default_process_types
     instrument "rails2.default_process_types" do
       web_process = bundler.has_gem?("thin") ?
-        "bundle exec thin start -e $RAILS_ENV -p $PORT" :
-        "bundle exec ruby script/server -p $PORT"
+        "bundle exec thin start -e $RAILS_ENV -p ${PORT:-5000}" :
+        "bundle exec ruby script/server -p ${PORT:-5000}"
 
       process_types = super
       process_types["web"]     = web_process
@@ -61,6 +61,12 @@ class LanguagePack::Rails2 < LanguagePack::Ruby
       install_plugins
       super
     end
+  end
+
+  def build
+    # TODO install plugins into separate layer
+    install_plugins
+    super
   end
 
   def best_practice_warnings
@@ -93,8 +99,8 @@ private
   end
 
   # sets up the profile.d script for this buildpack
-  def setup_profiled
-    super
+  def setup_profiled(*args)
+    super(*args)
     default_env_vars.each do |key, value|
       set_env_default key, value
     end
