@@ -12,10 +12,10 @@ module LanguagePack
       end
     end
 
-    DEFAULT_VERSION_NUMBER = "2.5.5"
-    DEFAULT_VERSION        = "ruby-#{DEFAULT_VERSION_NUMBER}"
-    LEGACY_VERSION_NUMBER  = "1.9.2"
-    LEGACY_VERSION         = "ruby-#{LEGACY_VERSION_NUMBER}"
+    DEFAULT_VERSION_NUMBER = "2.6.6".freeze
+    DEFAULT_VERSION        = "ruby-#{DEFAULT_VERSION_NUMBER}".freeze
+    LEGACY_VERSION_NUMBER  = "1.9.2".freeze
+    LEGACY_VERSION         = "ruby-#{LEGACY_VERSION_NUMBER}".freeze
     RUBY_VERSION_REGEX     = %r{
         (?<ruby_version>\d+\.\d+\.\d+){0}
         (?<patchlevel>p-?\d+){0}
@@ -38,6 +38,17 @@ module LanguagePack
       @version_without_patchlevel = @version.sub(/-p-?\d+/, '')
     end
 
+    def warn_ruby_26_bundler?
+      return false if Gem::Version.new(self.ruby_version) >= Gem::Version.new("2.6.3")
+      return false if Gem::Version.new(self.ruby_version) < Gem::Version.new("2.6.0")
+
+      return true
+    end
+
+    def ruby_192_or_lower?
+      Gem::Version.new(self.ruby_version) <= Gem::Version.new("1.9.2")
+    end
+
     # https://github.com/bundler/bundler/issues/4621
     def version_for_download
       if rbx?
@@ -47,6 +58,12 @@ module LanguagePack
       else
         version_without_patchlevel
       end
+    end
+
+    def file_name
+      file = "#{version_for_download}.tgz"
+      file.sub!("ruby", "ruby-build") if build?
+      file
     end
 
     # Before Ruby 2.1 patch releases were done via patchlevel i.e. 1.9.3-p426 versus 1.9.3-p448
@@ -94,6 +111,33 @@ module LanguagePack
     # does this vendor bundler
     def vendored_bundler?
       false
+    end
+
+    # Returns the next logical version in the minor series
+    # for example if the current ruby version is
+    # `ruby-2.3.1` then then `next_logical_version(1)`
+    # will produce `ruby-2.3.2`.
+    def next_logical_version(increment = 1)
+      return false if patchlevel_is_significant?
+      split_version = @version_without_patchlevel.split(".")
+      teeny = split_version.pop
+      split_version << teeny.to_i + increment
+      split_version.join(".")
+    end
+
+    def next_minor_version(increment = 1)
+      split_version = @version_without_patchlevel.split(".")
+      split_version[1] = split_version[1].to_i + increment
+      split_version[2] = 0
+      split_version.join(".")
+    end
+
+    def next_major_version(increment = 1)
+      split_version = @version_without_patchlevel.split("-").last.split(".")
+      split_version[0] = Integer(split_version[0]) + increment
+      split_version[1] = 0
+      split_version[2] = 0
+      return "ruby-#{split_version.join(".")}"
     end
 
     private
