@@ -1,51 +1,28 @@
 require_relative '../spec_helper'
 
 describe "Rails 3.x" do
-  it "should deploy on ruby 1.9.3" do
-    app = Hatchet::Runner.new("rails3_mri_193", stack: "cedar-14")
-    app.setup!
-    app.deploy do |app, heroku|
-      expect(app.output).to include("Asset precompilation completed")
+  it "should deploy and inject plugins" do
+    skip("Need RAILS_LTS_CREDS env var set") unless ENV["RAILS_LTS_CREDS"]
 
-      expect(app.output).to match("WARNING")
-      expect(app.output).to match("Add 'rails_12factor' gem to your Gemfile to skip plugin injection")
+    Hatchet::Runner.new("rails3_default_ruby", config: rails_lts_config).tap do |app|
+      app.before_deploy do
+        Pathname("Gemfile").write("ruby '2.7.2'", mode: "a")
+      end
 
-      ls = app.run("ls vendor/plugins")
-      expect(ls).to match("rails3_serve_static_assets")
-      expect(ls).to match("rails_log_stdout")
-    end
-  end
+      app.deploy do
+        # Rails 3 doesn't work with Postgres 8+ out of the box and Rails
+        # LTS hasn't patched this yet. We're skipping asset compilation for now
+        # by deleting the Rakefile
+        #
+        # expect(app.output).to include("Asset precompilation completed")
 
-  it "should not have warnings when using the rails_12factor gem" do
-    Hatchet::Runner.new("rails3_12factor").deploy do |app, heroku|
-      expect(app.output).not_to match("Add 'rails_12factor' gem to your Gemfile to skip plugin injection")
+        expect(app.output).to match("WARNING")
+        expect(app.output).to match("Add 'rails_12factor' gem to your Gemfile to skip plugin injection")
 
-      # https://github.com/heroku/heroku-buildpack-ruby/issues/525
-      env = app.run("env")
-      expect(env).not_to match("RAILS_GROUPS")
-    end
-  end
-
-  it "should only display the correct plugin warning" do
-    Hatchet::Runner.new("rails3_one_plugin").deploy do |app, heroku|
-      expect(app.output).not_to match("rails_log_stdout")
-      expect(app.output).to match("rails3_serve_static_assets")
-      expect(app.output).to match("Add 'rails_12factor' gem to your Gemfile to skip plugin injection")
-    end
-  end
-
-
-  it "fails if rake tasks cannot be detected" do
-    Hatchet::Runner.new("rails3-fail-rakefile", allow_failure: true).deploy do |app|
-      expect(app.output).to include("raising so the rake task will not load")
-      expect(app).not_to be_deployed
-    end
-  end
-
-  it "fails compile if assets:precompile fails" do
-    Hatchet::Runner.new("rails3-fail-assets-compile", allow_failure: true).deploy do |app, heroku|
-      expect(app.output).to include("raising on assets:precompile on purpose")
-      expect(app).not_to be_deployed
+        ls = app.run("ls vendor/plugins")
+        expect(ls).to match("rails3_serve_static_assets")
+        expect(ls).to match("rails_log_stdout")
+      end
     end
   end
 end
