@@ -1,10 +1,124 @@
 require_relative '../spec_helper'
 
 describe "Ruby apps" do
+
+  # https://github.com/heroku/heroku-buildpack-ruby/issues/1025
+  describe "bin/rake binstub" do
+    it "loads git gems at build time when executing `rake`" do
+      Hatchet::Runner.new("git_gemspec").tap do |app|
+        app.before_deploy do
+          File.open("Rakefile", "w+") do |f|
+            f.puts(<<~EOF)
+            task "assets:precompile" do
+              require 'mini_histogram'
+              puts "successfully loaded git gem"
+            end
+            EOF
+          end
+        end
+        app.deploy do
+          expect(app.output).to match("successfully loaded git gem")
+          expect(app.run("rake assets:precompile")).to match("successfully loaded git gem")
+        end
+      end
+    end
+
+    it "loads bundler into memory" do
+      Hatchet::Runner.new("default_ruby").tap do |app|
+        app.before_deploy do
+          File.open("Rakefile", "w+") do |f|
+            f.puts(<<~EOF)
+            task "assets:precompile" do
+              puts Bundler.methods
+
+              puts "bundler loaded in rake context"
+            end
+            EOF
+          end
+        end
+        app.deploy do
+          expect(app.output).to match("bundler loaded in rake context")
+          expect(app.run("rake assets:precompile")).to match("bundler loaded in rake context")
+        end
+      end
+    end
+
+    it "loads custom rake binstub" do
+      Hatchet::Runner.new("default_ruby").tap do |app|
+        app.before_deploy do
+          FileUtils.mkdir_p("bin")
+
+          File.open("bin/rake", "w+") do |f|
+            f.puts(<<~EOF)
+            #!/usr/bin/env ruby
+
+            puts "rake assets:precompile" # Needed to trigger the `rake -P` task detection
+            puts "custom rake binstub called"
+            EOF
+          end
+          FileUtils.chmod("+x", "bin/rake")
+        end
+        app.deploy do
+          expect(app.output).to match("custom rake binstub called")
+          expect(app.run("rake")).to match("custom rake binstub called")
+        end
+      end
+    end
+  end
+
+  describe "bad ruby version" do
+    it "gives a helpful error" do
+      Hatchet::Runner.new('ruby_version_does_not_exist', allow_failure: true, stack: DEFAULT_STACK).deploy do |app|
+        expect(app.output).to match("The Ruby version you are trying to install does not exist: ruby-2.9.0.lol")
+      end
+    end
+  end
+
+  describe "exporting path" do
+    it "puts local bin dir in path" do
+      before_deploy = Proc.new do
+        FileUtils.mkdir_p("bin")
+        File.open("bin/bloop", "w+") do |f|
+          f.puts(<<~EOF)
+          #!/usr/bin/env bash
+
+          echo "bloop"
+          EOF
+        end
+        FileUtils.chmod("+x", "bin/bloop")
+
+        File.open("Rakefile", "a") do |f|
+          f.puts(<<~EOF)
+          task "run:bloop" do
+            puts `bloop`
+            raise "Could not bloop" unless $?.success?
+          end
+          EOF
+        end
+      end
+      buildpacks = [
+        :default,
+        "https://github.com/schneems/buildpack-ruby-rake-deploy-tasks"
+      ]
+      config = { "DEPLOY_TASKS" => "run:bloop"}
+      Hatchet::Runner.new('default_ruby', stack: DEFAULT_STACK, buildpacks: buildpacks, config: config, before_deploy: before_deploy).deploy do |app|
+        expect(app.output).to match("bloop")
+      end
+    end
+  end
+
   describe "running Ruby from outside the default dir" do
     it "works" do
-      Hatchet::Runner.new('cd_ruby', stack: DEFAULT_STACK).deploy do |app|
+      buildpacks = [
+        :default,
+        "https://github.com/sharpstone/force_absolute_paths_buildpack"
+      ]
+      config = {FORCE_ABSOLUTE_PATHS_BUILDPACK_IGNORE_PATHS: "BUNDLE_PATH"}
+
+      Hatchet::Runner.new('cd_ruby', stack: DEFAULT_STACK, buildpacks: buildpacks, config: config).deploy do |app|
         expect(app.output).to match("cd version ruby 2.5.1")
+
+        expect(app.run("which ruby").strip).to eq("/app/bin/ruby")
       end
     end
   end
@@ -26,38 +140,13 @@ describe "Ruby apps" do
 
   describe "2.5.0" do
     it "works" do
-      Hatchet::Runner.new("ruby_25").deploy do
-        # works
+      Hatchet::Runner.new("ruby_25", stack: "heroku-18").deploy do |app|
+        expect(app.output).to include("There is a more recent Ruby version available")
       end
     end
   end
 
-  # describe "default WEB_CONCURRENCY" do
-  #   it "auto scales WEB_CONCURRENCY" do
-  #     pending("https://github.com/heroku/api/issues/4426")
-  #     app = Hatchet::Runner.new("default_ruby")
-  #     app.setup!
-  #     app.set_config("SENSIBLE_DEFAULTS" => "enabled")
-  #     app.deploy do |app|
-  #       app.run('echo "loaded"')
-  #       expect(app.run(:bash, 'echo "value: $WEB_CONCURRENCY"', heroku: { size: "1X" } )).to match("value: 2")
-  #       expect(app.run(:bash, 'echo "value: $WEB_CONCURRENCY"', heroku: { size: "2X" } )).to match("value: 4")
-  #       expect(app.run(:bash, 'echo "value: $WEB_CONCURRENCY"', heroku: { size: "PX" } )).to match("value: 16")
-  #     end
-  #   end
-  # end
-
   describe "Rake detection" do
-    context "default" do
-      # it "adds default process types" do
-      #   Hatchet::Runner.new('empty-procfile').deploy do |app|
-      #     app.run("console") do |console|
-      #       console.run("puts 'hello' + 'world'") {|result| expect(result).to match('helloworld')}
-      #     end
-      #   end
-      # end
-    end
-
     context "Ruby 1.9+" do
       it "runs a rake task if the gem exists" do
         Hatchet::Runner.new('default_with_rakefile').deploy do |app, heroku|
@@ -73,6 +162,7 @@ describe "Ruby apps" do
         Hatchet::Runner.new("default_ruby").deploy do |app, heroku|
           expect(app.output).to     include("Writing config/database.yml to read from DATABASE_URL")
           expect(app.output).not_to include("Your app was upgraded to bundler")
+          expect(app.output).not_to include("Your Ruby version is not present on the next stack")
         end
       end
     end
@@ -88,16 +178,15 @@ describe "Ruby apps" do
 end
 
 describe "Raise errors on specific gems" do
-  it "should should raise on sqlite3" do
-    Hatchet::Runner.new("sqlite3_gemfile", allow_failure: true).deploy do |app|
+  it "should raise on sqlite3" do
+    before_deploy = -> { run!(%Q{echo "ruby '2.5.4' >> Gemfile"}) }
+    Hatchet::Runner.new("sqlite3_gemfile", allow_failure: true, before_deploy: before_deploy).deploy do |app|
       expect(app).not_to be_deployed
       expect(app.output).to include("Detected sqlite3 gem which is not supported")
       expect(app.output).to include("devcenter.heroku.com/articles/sqlite3")
     end
   end
 end
-
-
 
 describe "No Lockfile" do
   it "should not deploy" do
@@ -111,10 +200,7 @@ end
 describe "Rack" do
   it "should not overwrite already set environment variables" do
     custom_env = "FFFUUUUUUU"
-    app = Hatchet::Runner.new("default_ruby")
-    app.setup!
-    app.set_config("RACK_ENV" => custom_env)
-    expect(app.run("env")).to match(custom_env)
+    app = Hatchet::Runner.new("default_ruby", config: {"RACK_ENV" => custom_env})
 
     app.deploy do |app|
       expect(app.run("env")).to match(custom_env)
